use Sourcery to generate MasterRuleList.swift
